Create the basic information for admin dashboard
I created the elements and sorted the information into different
containers/sections of the webpage. I commented each container and its
children what type of display will be needed to create the desired
layout. All the specific details and images will come later after I
create the layout as it is supposed to look like.